docs: обновить документацию
